Enable pipelining when supported
Enabling pipelining in Ansible reduces the number of SSH connections
opened for each task performed. It is not enabled by default because
it is incompatible with the requiretty sudoers setting that is found
in some Linux distributions' default configuration.

With this change, we enable pipelining at the node level if we detect
that the sudoers file does not have the requiretty setting.