Use semver library and fix version file checking
Given that we are going to perform pre-releases, we decided to use a
supported and tested semantic version library instead of trying to
implement our own. Version detection is important as it is in the
upgrade path.

Also fixed an issue where we were assuming v1.0.0 if any error ocurred
while attempting to read the version file.