Use official calico CNI installation script
Historically, we have controlled the deployment of the calico policy controller using a flag in the plan file. Now that we deploy the controller out of the box, and we removed the flag, there is no need to use a custom calico CNI configuration file. For this reason, we can rely on the upstream installation script maintained by the calico team.

To avoid having to create a secret for a certificate that is already on the node, we configure the location of the etcd certificates manually (in the cni_network_config), instead of relying on the template support provided by the calico script.