Fixing diagnose command
We were seeing in integration tests that the diagnose command was
getting stuck. After investigation, I found that the fetch module was
logging the contents of the diagnose tarball. This, in turn, was causing
an error in the scanner we use for timestamping the ansible log file, as
the contents of the file were too large to fit in the scanner's buffer.

The fetch module has a warning around this, and suggests setting become
= False to avoid this issue.