Set hostname to lower when using nodenames and static pods
The kubelet uses the nodes hostname in a few places, however it sets any uppercase to lowercase. We need to follow the same strategy when interacting with the cluster. Specifically the nodename and the name of a static pod.

Also modified an existing integration test to cover this scenario.

Will need to add an upgrade test in the next release to test this scenario.