Skip drain if API server is not running
When upgrading master nodes, there is the possibility of entering a
scenario where the API server has been stopped, but the upgrade has to
be retried due to failure. For this reason, specifically in master
nodes, the drain command only runs if the API server is supposed to be
running.