Rely on existing node label kubernetes.io/hostname
In KET 1.6.x, the installer labeled nodes with kismatic/host= using the node-labels flag on the kubelet. However that flag only gets used on initial node registration. To properly support upgrades from previous versions and to be able to perform node selection, we can rely on an existing label kubernetes.io/hostname that is always set by kubernetes.