Enable PackageNotInstalled inspector tests
Note, these tests will not run if the user has specified docker.disable.
These tests were originally disabled because it did not support a use case when the user has already installed docker and manually commented out the installation steps. With the introduction of the new plan file option, the user will not need to manually modify the Ansible plays and it is safe to run these now.